Check permissions.json before applying permissions Some cleanup

+++ includes/prehtml/internalarea/usermanager.php
@@ -16,6 +16,7 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 					}
 				}
 			}
+
 			function applypermissions_searchUserInGroups($groupList, $groups, $userId, &$foundGroups)
 			{
 				foreach ($groups as $group)
@@ -34,6 +35,7 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 					}
 				}
 			}
+
 			function applypermissions_setGroupIds(&$groups, &$groupList, $id = 0)
 			{
 				$parentGroupId = $id;
@@ -55,64 +57,63 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 						$groupList[$groupId]->subGroups = $childIds;
 					}
 				}
+
 				return $id;
 			}
-			
-			$sourceData = json_decode(file_get_contents(ROOT_PATH . "/includes/permissions.json"));
-			$groupList = array();
-			applypermissions_setGroupIds($sourceData, $groupList);
-			
+
 			$errors = 0;
 			$ok = 0;
-			Constants::$pdo->query("TRUNCATE TABLE `permissions`");
-			$query = Constants::$pdo->query("SELECT `id` FROM `users`");
-			$permissionQuery = Constants::$pdo->prepare("INSERT INTO `permissions` (`userId`, `permission`) VALUES(:userId, :permission)");
-			while ($row = $query->fetch())
+
+			$sourceData = json_decode(file_get_contents(ROOT_PATH . "/includes/permissions.json"));
+			if ($sourceData)
 			{
-				$users++;
-				$foundGroups = array();
-				$permissions = array();
-				applypermissions_searchUserInGroups($groupList, $sourceData, $row->id, $foundGroups);
-				foreach ($foundGroups as $groupId)
+				$groupList = array();
+				applypermissions_setGroupIds($sourceData, $groupList);
+
+				Constants::$pdo->query("TRUNCATE TABLE `permissions`");
+				$query = Constants::$pdo->query("SELECT `id` FROM `users`");
+				$permissionQuery = Constants::$pdo->prepare("INSERT INTO `permissions` (`userId`, `permission`) VALUES(:userId, :permission)");
+				while ($row = $query->fetch())
 				{
-					$group = $groupList[$groupId];
-					if ($group->permissions)
+					$users++;
+					$foundGroups = array();
+					$permissions = array();
+					applypermissions_searchUserInGroups($groupList, $sourceData, $row->id, $foundGroups);
+
+					foreach ($foundGroups as $groupId)
 					{
-						foreach ($group->permissions as $permission)
+						$group = $groupList[$groupId];
+						if ($group->permissions)
 						{
-							if (!in_array($permission, $permissions))
+							foreach ($group->permissions as $permission)
 							{
-								$permissions[] = $permission;
+								if (!in_array($permission, $permissions))
+								{
+									$permissions[] = $permission;
+								}
 							}
 						}
 					}
-				}
-				foreach ($permissions as $permission)
-				{
-					$queryData = array
-					(
-						":userId" => $row->id,
-						":permission" => $permission
-					);
-					if ($permissionQuery->execute($queryData))
+
+					foreach ($permissions as $permission)
 					{
-						$ok++;
-					}
-					else
-					{
-						$errors++;
+						$queryData = array(":userId" => $row->id, ":permission" => $permission);
+						if ($permissionQuery->execute($queryData))
+						{
+							$ok++;
+						} else
+						{
+							$errors++;
+						}
 					}
 				}
 			}
-			
-			echo json_encode(array
-			(
-				"errors" => $errors,
-				"ok" => $ok
-			));
+
+			echo json_encode(array("errors" => $errors, "ok" => $ok));
 			exit;
 		case "getpermissiongroups":
 			$sourceData = json_decode(file_get_contents(ROOT_PATH . "/includes/permissions.json"));
+
 			function getpermissiongroups_createTree($sourceData)
 			{
 				$data = array();
@@ -121,10 +122,12 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 					$groupData = new StdClass;
 					$groupData->data = $group->title;
 					$children = array();
+
 					if ($group->subGroups and !empty($group->subGroups))
 					{
 						$children = getpermissiongroups_createTree($group->subGroups);
 					}
+
 					if ($group->permissions and !empty($group->permissions))
 					{
 						$permissions = array();
@@ -136,94 +139,64 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 								$type = "permission_revoked";
 								$permission = substr($permission, 1);
 							}
-							$permissions[] = array
-							(
-								"data" => $permission,
-								"attr" => array
-								(
-									"rel" => $type
-								)
-							);
+							$permissions[] = array("data" => $permission, "attr" => array("rel" => $type));
 						}
 						if (!empty($permissions))
 						{
-							$children[] = array
-							(
-								"data" => "Berechtigungen",
-								"attr" => array
-								(
-									"rel" => "permission"
-								),
-								"children" => $permissions
-							);
+							$children[] = array("data" => "Berechtigungen", "attr" => array("rel" => "permission"), "children" => $permissions);
 						}
 					}
+
 					if ($group->users and !empty($group->users))
 					{
 						$users = array();
 						$query = Constants::$pdo->query("SELECT `id`, `firstName`, `lastName` FROM `users` WHERE `id` IN(" . implode(",", $group->users) . ") ORDER BY `lastName`, `firstName`");
 						while ($row = $query->fetch())
 						{
-							$users[] = array
-							(
-								"data" => $row->firstName . " " . $row->lastName,
-								"attr" => array
-								(
-									"rel" => "user"
-								)
-							);
+							$users[] = array("data" => $row->firstName . " " . $row->lastName, "attr" => array("rel" => "user"));
 						}
 						if (!empty($users))
 						{
-							$children[] = array
-							(
-								"data" => "Benutzer",
-								"attr" => array
-								(
-									"rel" => "user"
-								),
-								"children" => $users
-							);
+							$children[] = array("data" => "Benutzer", "attr" => array("rel" => "user"), "children" => $users);
 						}
 					}
+
 					if ($children)
 					{
 						$groupData->children = $children;
 					}
+
 					$data[] = $groupData;
 				}
+
 				return $data;
 			}
+
 			echo json_encode(getpermissiongroups_createTree($sourceData));
 			exit;
 		case "getuserdata":
 			$query = Constants::$pdo->prepare("SELECT `id`, `enabled`, `username`, `email`, `firstName`, `lastName`, `birthDate` FROM `users` WHERE `id` = :id");
-			$query->execute(array
-			(
-				":id" => Constants::$pagePath[3]
-			));
+			$query->execute(array(":id" => Constants::$pagePath[3]));
 			$row = $query->fetch();
-			$row->id = (int) $row->id;
-			$row->enabled = (bool) $row->enabled;
+			$row->id = (int)$row->id;
+			$row->enabled = (bool)$row->enabled;
 			$data = $row;
-			
+
 			$query = Constants::$pdo->prepare("SELECT `id`, `category`, `subCategory`, `number` FROM `phonenumbers` WHERE `userId` = :userId");
-			$query->execute(array
-			(
-				":userId" => $data->id
-			));
+			$query->execute(array(":userId" => $data->id));
 			$data->phoneNumbers = $query->fetchAll();
-			
+
 			$profilePictureFile = "/files/profilepictures/" . $data->id . ".jpg";
 			if (file_exists(ROOT_PATH . $profilePictureFile))
 			{
 				$data->profilePictureUrl = "/getprofilepicture/" . $data->id . "/" . md5_file(ROOT_PATH . $profilePictureFile);
 			}
-			
+
 			echo json_encode($data);
 			exit;
 		case "getuserpermissions":
 			$sourceData = json_decode(file_get_contents(ROOT_PATH . "/includes/permissions.json"));
+
 			function getuserpermissions_createTree($sourceData)
 			{
 				$data = array();
@@ -231,14 +204,8 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 				{
 					$groupData = new StdClass;
 					$groupData->data = $group->title;
-					$groupData->attr = array
-					(
-						"class" => ($group->users and !empty($group->users) and in_array(Constants::$pagePath[3], $group->users)) ? "jstree-checked" : ""
-					);
-					$groupData->metadata = array
-					(
-						"groupId" => $group->id
-					);
+					$groupData->attr = array("class" => ($group->users and !empty($group->users) and in_array(Constants::$pagePath[3], $group->users)) ? "jstree-checked" : "");
+					$groupData->metadata = array("groupId" => $group->id);
 					$children = array();
 					if ($group->subGroups and !empty($group->subGroups))
 					{
@@ -246,8 +213,10 @@ if (Constants::$accountManager->hasPermission("usermanager"))
 					}
 					$data[] = $groupData;
 				}
+
 				return $data;
 			}
+
 			echo json_encode(getuserpermissions_createTree($sourceData));
 			exit;
 	}
